pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kepler-framework"
version = "0.1.0"
description = "Simple framework for industrial ML with Splunk and Google Cloud"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Kepler Team", email = "kepler@company.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.8"
dependencies = [
    # CLI framework
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    
    # Data processing
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    
    # ML frameworks (MVP: only sklearn + xgboost)
    "scikit-learn>=1.2.0",
    "xgboost>=1.7.0",
    "joblib>=1.2.0",
    
    # Splunk integration
    "splunk-sdk>=1.7.0",
    "requests>=2.28.0",
    
    # GCP integration (basic)
    "google-cloud-run>=0.10.0",
    "google-cloud-storage>=2.7.0",
    "docker>=6.0.0",
    
    # Configuration management
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "pathlib>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
kepler = "kepler.cli:app"

[project.urls]
Homepage = "https://github.com/company/kepler-framework"
Repository = "https://github.com/company/kepler-framework"
Documentation = "https://kepler-framework.readthedocs.io"

[tool.setuptools.packages.find]
where = ["."]
include = ["kepler*"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=kepler --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # MVP: relaxed typing
ignore_missing_imports = true  # MVP: don't be strict about external types